pyproject.toml
[project]
name = "yazis"
version = "0.1.0"
description = "Natural Language Dictionary Formation System for English text processing"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "PyQt6",
    "python-docx",
    "nltk",
    "spacy",
    "pandas",
    "pyyaml",
    "groq",
]

[project.scripts]
yazis = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/config", "src/core", "src/data", "src/gui", "src/models", "src/utils"]

[tool.hatch.build.targets.wheel.force-include]
"src/main.py" = "main.py"
"src/resources" = "resources"

[dependency-groups]
dev = [
    "pytest",
]
